refactored tracks viewModel to call repository to get audiofiles

+++ app/src/main/java/com/rickinc/decibels/presentation/model/Track.kt
@@ -1,9 +1,9 @@
 package com.rickinc.decibels.presentation.model
 
 data class Track(
-    val trackId: Int,
+    val trackId: Long,
     val trackName: String,
-    val trackLength: Long
+    val trackLength: Int
 ) {
     companion object {
         fun createDummyTracks(): List<Track> {
